119143 - invalidate odds cache

+++ src/Shared/Core/Models/Odds/BetTypeOdds.cs
@@ -1,14 +1,14 @@
 ﻿namespace LiveScore.Core.Models.Odds
 {
     using System.Collections.Generic;
-    public interface IBetTypeOdds : IEntity<int, string>
+    public interface IBetTypeOdds : IEntity<byte, string>
     {
         Bookmaker Bookmaker { get; }
 
         IEnumerable<BetOptionOdds> BetOptions { get; set; }
     }
 
-    public class BetTypeOdds : Entity<int, string>, IBetTypeOdds
+    public class BetTypeOdds : Entity<byte, string>, IBetTypeOdds
     {
         public Bookmaker Bookmaker { get; set; }
