+++ src/Shared/Core/Models/Odds/OddsEvent.cs
@@ -2,7 +2,7 @@
 {
     public interface IOddsMovementEvent
     {
-        int BetTypeId { get; }
+        byte BetTypeId { get; }
 
         Bookmaker Bookmaker { get; }
